Frequency Dependent source
